// src/presentation/navigation/types.ts
// Tipos de navegacion centralizados

export type RootStackParamList = {
  Welcome: undefined;
  Login: undefined;
  Register: undefined;
  Main: undefined;
  ComponentPreview: undefined;
};

export type TabParamList = {
  Home: undefined;
  Transactions: undefined;
  Budget: undefined;
  AssetsLiabilities: undefined;
  Profile: undefined;
};